refactor: unify ParsedFieldValue into IdlValue
Remove ParsedFieldValue entirely — use sonar_idl::IdlValue as the
single value type across both the IDL decoder and built-in parsers.

- Move Serialize, to_json_value(), PartialEq<&str> into IdlValue
- Enable arbitrary_precision in sonar-idl for u128/i128 JSON numbers
- ParsedField::value is now IdlValue directly
- Anchor adapter passes IdlValue through without conversion
- Account handler uses IdlValue::to_json_value() directly
- Eliminates the From<IdlValue> for ParsedFieldValue mapping layer
