+++ crates/sonar-idl/Cargo.toml
@@ -11,6 +11,6 @@ heck = "0.5"
 hex = "0.4"
 log = "0.4"
 serde = { version = "1.0", features = ["derive"] }
-serde_json = { version = "1.0", features = ["preserve_order"] }
+serde_json = { version = "1.0", features = ["preserve_order", "arbitrary_precision"] }
 sha2 = "0.10.9"
 solana-pubkey = "3.0"
