+++ src/parsers/instruction/types.rs
@@ -2,6 +2,7 @@ use std::ops::Deref;
 
 use serde::Serialize;
 use solana_pubkey::Pubkey;
+use sonar_idl::IdlValue;
 
 /// Represents a parsed instruction with human-readable data and account names
 #[derive(Debug, Clone, Serialize)]
@@ -77,28 +78,28 @@ impl Serialize for ParsedInstructionFields {
 #[derive(Debug, Clone, Serialize)]
 pub struct ParsedField {
     pub name: String,
-    pub value: ParsedFieldValue,
+    pub value: IdlValue,
 }
 
 impl ParsedField {
     pub fn text(name: impl Into<String>, value: impl Into<String>) -> Self {
-        Self { name: name.into(), value: ParsedFieldValue::Text(value.into()) }
+        Self { name: name.into(), value: IdlValue::String(value.into()) }
     }
 
     pub fn number(name: impl Into<String>, value: u64) -> Self {
-        Self { name: name.into(), value: ParsedFieldValue::U64(value) }
+        Self { name: name.into(), value: IdlValue::U64(value) }
     }
 
     pub fn signed_number(name: impl Into<String>, value: i64) -> Self {
-        Self { name: name.into(), value: ParsedFieldValue::I64(value) }
+        Self { name: name.into(), value: IdlValue::I64(value) }
     }
 
     pub fn boolean(name: impl Into<String>, value: bool) -> Self {
-        Self { name: name.into(), value: ParsedFieldValue::Bool(value) }
+        Self { name: name.into(), value: IdlValue::Bool(value) }
     }
 
     pub fn pubkey(name: impl Into<String>, value: Pubkey) -> Self {
-        Self { name: name.into(), value: ParsedFieldValue::Pubkey(value) }
+        Self { name: name.into(), value: IdlValue::Pubkey(value) }
     }
 }
 
@@ -109,121 +110,5 @@ where
 {
     fn from((name, value): (N, V)) -> Self {
         ParsedField::text(name, value)
-    }
-}
-
-/// Domain-native parsed field value mirroring the IDL/Solana type system.
-///
-/// Each variant preserves the exact type and bit width from the source,
-/// giving consumers full context for serialization and display decisions.
-#[derive(Debug, Clone, PartialEq)]
-pub enum ParsedFieldValue {
-    U8(u8),
-    U16(u16),
-    U32(u32),
-    U64(u64),
-    U128(u128),
-    I8(i8),
-    I16(i16),
-    I32(i32),
-    I64(i64),
-    I128(i128),
-    Bool(bool),
-    Pubkey(Pubkey),
-    Text(String),
-    Bytes(Vec<u8>),
-    /// Ordered named fields (struct, nested object).
-    Struct(Vec<(String, ParsedFieldValue)>),
-    /// Ordered values (vec, array).
-    Array(Vec<ParsedFieldValue>),
-    Null,
-}
-
-/// Serialize for `--json` output. Delegates entirely to `to_json_value()`.
-///
-/// With `arbitrary_precision`, u128/i128 emit as unquoted JSON numbers
-/// regardless of magnitude — no special-casing needed.
-impl Serialize for ParsedFieldValue {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        self.to_json_value().serialize(serializer)
-    }
-}
-
-impl ParsedFieldValue {
-    /// Convert to `serde_json::Value`.
-    ///
-    /// All integer types (including u128/i128) emit as JSON numbers via
-    /// `serde_json`'s `arbitrary_precision` feature.
-    pub fn to_json_value(&self) -> serde_json::Value {
-        use serde_json::{Number, Value};
-        match self {
-            Self::U8(n) => Value::Number((*n as u64).into()),
-            Self::U16(n) => Value::Number((*n as u64).into()),
-            Self::U32(n) => Value::Number((*n as u64).into()),
-            Self::U64(n) => Value::Number((*n).into()),
-            Self::U128(n) => Value::Number(Number::from_u128(*n).unwrap()),
-            Self::I8(n) => Value::Number((*n as i64).into()),
-            Self::I16(n) => Value::Number((*n as i64).into()),
-            Self::I32(n) => Value::Number((*n as i64).into()),
-            Self::I64(n) => Value::Number((*n).into()),
-            Self::I128(n) => Value::Number(Number::from_i128(*n).unwrap()),
-            Self::Bool(b) => Value::Bool(*b),
-            Self::Pubkey(p) => Value::String(p.to_string()),
-            Self::Text(s) => Value::String(s.clone()),
-            Self::Bytes(bytes) => {
-                Value::Array(bytes.iter().map(|b| Value::Number((*b as u64).into())).collect())
-            }
-            Self::Struct(fields) => {
-                let map: serde_json::Map<String, Value> =
-                    fields.iter().map(|(k, v)| (k.clone(), v.to_json_value())).collect();
-                Value::Object(map)
-            }
-            Self::Array(values) => {
-                Value::Array(values.iter().map(|v| v.to_json_value()).collect())
-            }
-            Self::Null => Value::Null,
-        }
-    }
-}
-
-impl From<String> for ParsedFieldValue {
-    fn from(value: String) -> Self {
-        ParsedFieldValue::Text(value)
-    }
-}
-
-impl From<&str> for ParsedFieldValue {
-    fn from(value: &str) -> Self {
-        ParsedFieldValue::Text(value.to_string())
-    }
-}
-
-impl PartialEq<&str> for ParsedFieldValue {
-    fn eq(&self, other: &&str) -> bool {
-        match self {
-            Self::Text(s) => s == *other,
-            Self::U8(n) => n.to_string() == *other,
-            Self::U16(n) => n.to_string() == *other,
-            Self::U32(n) => n.to_string() == *other,
-            Self::U64(n) => n.to_string() == *other,
-            Self::U128(n) => n.to_string() == *other,
-            Self::I8(n) => n.to_string() == *other,
-            Self::I16(n) => n.to_string() == *other,
-            Self::I32(n) => n.to_string() == *other,
-            Self::I64(n) => n.to_string() == *other,
-            Self::I128(n) => n.to_string() == *other,
-            Self::Bool(b) => b.to_string() == *other,
-            Self::Pubkey(p) => p.to_string() == *other,
-            _ => false,
-        }
-    }
-}
-
-impl PartialEq<String> for ParsedFieldValue {
-    fn eq(&self, other: &String) -> bool {
-        self == &other.as_str()
     }
 }
